skeleton project
This first archive is the skeleton project for Visual Studio Code. It just reads both json input files and parses those into Newtonsoft.Json arrays. Purpose is to find out how fast the input files can be read and parsed in memory, and to find out what exactly is inside that data.